+++ Code/Library/PathTools.h
@@ -4,11 +4,18 @@
 
 namespace PathTools
 {
+	bool IsAbsolute(StringView path);
+
 	StringView BaseName(StringView path);
 	StringView DirName(StringView path);
 
 	StringView GetFileExtension(StringView path);
 	StringView RemoveFileExtension(StringView path);
+
+	std::string GetWorkingDirectory();
+	std::string GetDocumentsPath();
+
+	std::string MakeAbsolute(StringView path);
 
 	std::string Join(StringView pathA, StringView pathB);
 }