+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- `-userdirname` to change name of user directory in `Documents/My Games/`. Overrides `Game/Config/Folders.ini`.
+    - `-userdirname Test` => `Documents/My Games/Test`
+    - `-userdirname "My New Mod"` => `Documents/My Games/My New Mod`
+- `-userpath` to completely change user directory path. Both `\` and `/` are supported. Overrides `-userdirname`.
+    - `-userpath .` => Crysis main directory
+    - `-userpath Something\MyFolder` => Crysis main directory + `Something\MyFolder` (relative path)
+    - `-userpath C:\Something\MyFolder` => `C:\Something\MyFolder` (absolute path)
 ### Fixed
 - Incorrectly trimmed spaces from CPU model name in specific cases.
 - Rare crash in the engine due to out-of-bounds access of the `CPUInfo::cores` array.